+++ CHANGELOG.md
@@ -3,9 +3,10 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [20.2.3] - (not yet released)
+## [20.2.3] - 2023-03-17
 - Audit.EntityFramework: Fix to honor Override/Ignore/Format in entity properties for any mapper type (#584)
 - Audit.EntityFramework: Fix EntityType null in the ExplicitMapper when using sync methods (#585)
+- Audit.EntityFramework: Fix Property Matcher for entities with custom column name (#586)
 
 ## [20.2.2] - 2023-03-14
 - Audit.EntityFramework / Audit.EntityFramework.Core: Low-Level interceptors support to use the `.AuditDataProvider` property from the `AuditDbContext` instance (#387)